requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-decouple==3.8
dj-database-url==2.3.0
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
sendgrid==6.11.0
razorpay==1.4.2
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0
django-filter==24.3
Pillow==11.1.0
whitenoise==6.8.2
